requirements.txt
# ========== DEPENDÊNCIAS PRINCIPAIS ==========
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-OAuthlib==0.9.6
Flask-Caching==2.3.1
Flask-Limiter==4.0.0
Flask-WTF==1.2.2
Flask-JWT-Extended==4.7.1
Flask-CORS==4.0.0

# ========== DEPENDÊNCIAS DE SEGURANÇA ==========
PyJWT==2.10.1
Werkzeug==2.3.7
bleach==6.2.0
markupsafe==3.0.3
cryptography==41.0.7
bcrypt==4.1.2
passlib==1.7.4

# ========== DEPENDÊNCIAS DE AUTENTICAÇÃO ==========
pyotp==2.9.0
qrcode[pil]==8.2
oauthlib==2.1.0
requests-oauthlib==1.1.0

# ========== DEPENDÊNCIAS DE CACHE E STORAGE ==========
cachelib==0.13.0
redis==6.4.0

# ========== DEPENDÊNCIAS DE UTILITÁRIOS ==========
python-dotenv==1.0.0
reportlab==4.0.4
openpyxl==3.1.2
requests==2.31.0
psycopg2-binary==2.9.9

# ========== DEPENDÊNCIAS DE MONITORAMENTO ==========
sentry-sdk[flask]==1.38.0
prometheus-client==0.19.0

# ========== DEPENDÊNCIAS DE TESTE ==========
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0